+++ src/components/drag-and-drop/with-drag/__spec__.js
@@ -1,32 +1,123 @@
-// __spec__.js
 import React from 'react';
+import PropTypes from 'prop-types';
 import { mount } from 'enzyme';
-
-import DraggableContext from './../draggable-context';
+import { DragDropContext } from 'react-dnd';
+import TestBackend from 'react-dnd-test-backend';
 import WithDrag from './with-drag';
-import { DragSource } from 'react-dnd';
 
-fdescribe('WithDrag', () => {
-  let wrapper;
+describe('WithDrag', () => {
+  let wrapper, backend, handlerId, beginDragContextSpy, beginDragPropSpy,
+      endDragContextSpy, endDragPropSpy;
+
+  function createWrapper(props = {}) {
+    let DnD = wrapInTestContext(WithDrag);
+
+    wrapper = mount(
+      <DnD { ...props }>
+        <div>foo</div>
+      </DnD>
+    );
+
+    let component = wrapper.find(WithDrag).getNode();
+    handlerId = component.getHandlerId();
+    backend = wrapper.getNode().getManager().backend;
+  }
+
+  function wrapInTestContext(DecoratedComponent) {
+    return DragDropContext(TestBackend)(
+      class TestContextContainer extends React.Component {
+        static childContextTypes = {
+          dragAndDropBeginDrag: PropTypes.func,
+          dragAndDropEndDrag: PropTypes.func
+        }
+
+        getChildContext() {
+          return {
+            dragAndDropBeginDrag: beginDragContextSpy,
+            dragAndDropEndDrag: endDragContextSpy,
+          };
+        }
+
+        render() {
+          return <DecoratedComponent {...this.props} />;
+        }
+      }
+    );
+  }
 
   beforeEach(() => {
-    wrapper = mount(
-      <DraggableContext>
-        <div>
-          <WithDrag>
-            <div>foo</div>
-          </WithDrag>
-        </div>
-      </DraggableContext>
-    );
+    beginDragContextSpy = jasmine.createSpy().and.returnValue({
+      index: 1
+    });
+    beginDragPropSpy = jasmine.createSpy().and.returnValue({
+      index: 1
+    });
+    endDragContextSpy = jasmine.createSpy();
+    endDragPropSpy = jasmine.createSpy();
   });
 
-  describe('canDrag', () => {
-    describe('with no custom method defined', () => {
-      it('returns true', () => {
-        let w = DragSource;
-        debugger
-        expect(wrapper.canDrag()).toEqual(true);
+  describe('without custom props', () => {
+    beforeEach(() => {
+      createWrapper();
+    });
+
+    describe('beginDrag', () => {
+      it('calls the beginDrag from context', () => {
+        backend.simulateBeginDrag([handlerId]);
+        expect(beginDragPropSpy).not.toHaveBeenCalled();
+        expect(beginDragContextSpy).toHaveBeenCalled();
+      });
+    });
+
+    describe('endDrag', () => {
+      it('calls the endDrag from context', () => {
+        backend.simulateBeginDrag([handlerId]);
+        backend.simulateEndDrag([handlerId]);
+        expect(endDragPropSpy).not.toHaveBeenCalled();
+        expect(endDragContextSpy).toHaveBeenCalled();
+      });
+    });
+  });
+
+  describe('with custom props', () => {
+    beforeEach(() => {
+      createWrapper({
+        beginDrag: beginDragPropSpy,
+        endDrag: endDragPropSpy
+      });
+    });
+
+    describe('beginDrag', () => {
+      it('calls the beginDrag from context', () => {
+        backend.simulateBeginDrag([handlerId]);
+        expect(beginDragPropSpy).toHaveBeenCalled();
+        expect(beginDragContextSpy).not.toHaveBeenCalled();
+      });
+    });
+
+    describe('endDrag', () => {
+      it('calls the endDrag from context', () => {
+        backend.simulateBeginDrag([handlerId]);
+        backend.simulateEndDrag([handlerId]);
+        expect(endDragPropSpy).toHaveBeenCalled();
+        expect(endDragContextSpy).not.toHaveBeenCalled();
+      });
+    });
+  });
+
+  describe('custom canDrag', () => {
+    beforeEach(() => {
+      createWrapper({
+        canDrag: () => {
+          return false
+        }
+      });
+    });
+
+    describe('when it returns false', () => {
+      it('it does not trigger begin drag', () => {
+        backend.simulateBeginDrag([handlerId]);
+        expect(beginDragContextSpy).not.toHaveBeenCalled();
       });
     });
   });