+++ src/components/table/draggable-table-cell/draggable-table-cell.js
@@ -2,6 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import WithDrag from './../../drag-and-drop/with-drag';
 import Icon from './../../icon';
+import TableCell from './../table-cell';
 
 /**
  * Creates a draggable table cell using WithDrag.