+++ src/utils/helpers/dnd/item-target/__spec__.js
@@ -3,14 +3,13 @@ import ReactDOM from 'react-dom';
 import ItemTargetHelper from './';
 
 describe('ItemTargetHelper', () => {
-
   describe('onHoverUpDown', () => {
-    let component;
-    let rect;
-    let monitor = {};
-    let monitorItem;
-    let props;
-    let clientOffset;
+    let component,
+        rect,
+        monitor = {},
+        monitorItem,
+        props,
+        clientOffset;
 
     beforeEach(() => {
       component = document.createElement('div');
@@ -25,7 +24,7 @@ describe('ItemTargetHelper', () => {
 
       props = {
         index: 0,
-        moveItem: () => {}
+        onDrag: () => {}
       };
       
       clientOffset = {
@@ -50,15 +49,15 @@ describe('ItemTargetHelper', () => {
       };
     });
 
-    describe('calls props.moveItem when', () => {
+    describe('calls props.onDrag when', () => {
       it('item dragged downwards and below 50% of item height', () => {
         props.index = 1;
         clientOffset.y = 76; // over the halfway point when dragging downwards
 
-        spyOn(props, 'moveItem');
+        spyOn(props, 'onDrag');
         ItemTargetHelper.onHoverUpDown(props, monitor, component);
 
-        expect(props.moveItem).toHaveBeenCalled();
+        expect(props.onDrag).toHaveBeenCalled();
       });
 
       it('item dragged upwards and above 50% of item height', () => {
@@ -67,32 +66,48 @@ describe('ItemTargetHelper', () => {
         clientOffset.y = 51;
         monitorItem.index = 1;
 
-        spyOn(props, 'moveItem');
+        spyOn(props, 'onDrag');
         ItemTargetHelper.onHoverUpDown(props, monitor, component);
 
-        expect(props.moveItem).toHaveBeenCalled();
+        expect(props.onDrag).toHaveBeenCalled();
+      });
+
+      it('calls context.dragAndDropOnDrag when onDrag prop is unavailable is', () => {
+        props.index = 0;
+
+        clientOffset.y = 51;
+        monitorItem.index = 1;
+        let contextSpy = jasmine.createSpy();
+
+        component.context = {
+          dragAndDropOnDrag: contextSpy
+        };
+        delete props.onDrag;
+
+        ItemTargetHelper.onHoverUpDown(props, monitor, component);
+        expect(contextSpy).toHaveBeenCalled();
       });
     });
 
-    describe('does not call props.moveItem when', () => {
+    describe('does not call props.onDrag when', () => {
       it('dragIndex is the same as the hoverIndex', () => {
         props.index = 0;
         monitorItem.index = 0;
 
-        spyOn(props, 'moveItem');
+        spyOn(props, 'onDrag');
 
         ItemTargetHelper.onHoverUpDown(props, monitor, component);
-        expect(props.moveItem).not.toHaveBeenCalled();
+        expect(props.onDrag).not.toHaveBeenCalled();
       });
 
       it('dragging down and not above 50% of the item height', () => {
         props.index = 1;
         clientOffset.y = 70;
 
-        spyOn(props, 'moveItem');
+        spyOn(props, 'onDrag');
         ItemTargetHelper.onHoverUpDown(props, monitor, component);
 
-        expect(props.moveItem).not.toHaveBeenCalled();
+        expect(props.onDrag).not.toHaveBeenCalled();
       });
 
       it('dragging up and not above 50% of the item height', () => {
@@ -101,12 +116,11 @@ describe('ItemTargetHelper', () => {
         clientOffset.y = 76;
         monitorItem.index = 1;
 
-        spyOn(props, 'moveItem');
+        spyOn(props, 'onDrag');
         ItemTargetHelper.onHoverUpDown(props, monitor, component);
 
-        expect(props.moveItem).not.toHaveBeenCalled();
+        expect(props.onDrag).not.toHaveBeenCalled();
       });
     });
   });
-  
 });